Updated to work with Blender 2.93.1
This is technically my first work on a Blender addon so hopefully it comes in handy for someone else. Just had to update some syntax on labels but I think the file browser on the export path button is still broken. Unfortunately I'm not knowledgable to fix that one :S